requirements.txt
httpx
openai-whisper
openai-whisper
chromadb
pysqlite3-binary
